+++ client/src/components/Title.css
@@ -4,9 +4,8 @@
   display: flex;
   justify-content: center;
   align-items: baseline;
-  padding: 10em;
-  margin-bottom: 10em;
-  background-image: url(./My-Lit-Life_Title-Img.png) !important;
+  padding: 5em;
+  margin-bottom: 2em;
   background-color: rgb(243, 227, 81);
   color: rgb(240, 137, 20);
 }
