--- client/src/components/pages/Title.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { useAuth } from '../authentication/context/AuthContext';
-import './Title.css';
-
-function Title() {
-    const { currentUser } = useAuth();
-    console.log(currentUser);
-
-  return (
-    <div className="Title">
-      <h1 className="display-4">Hello, {currentUser.displayName.split(' ')[0]}</h1>
-      </div>
-  )
-}
-
-export default Title;
